Add files via upload
This version has a unique line for every fragment in every condition. There are issues with two fragments: the planes l_nd(land/step) and the trains st_p(stop/step). Not sure what to do here. I can't think of other words which conform to the requirements. Leave for now.