Add files via upload
These gestures are without audio and with the face covered. The clips with two syllable nouns (the baby's/babies and horse's/horses) are 2.1 (for the + noun) + 1.5 seconds (for the POS, PL or general S). The clips with one syllable nouns (the cat's/cats, the dog's/dogs etc.) are 1.8 + 1.5 (for the POS, PL or general S) seconds long.